Isolate and improve Astropy table data factory
