+++ glue/core/data_factories/__init__.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, division, print_function
 from .helpers import *
 from .fits import *
 from .hdf5 import *
+from .astropy_table import *
 from .pandas import *
 from .excel import *
 from .image import *
